Add files via upload
This function is used to download the Himawari-8 remote images from the 'ftpsite: ftp.ptree.jaxa.jp'. Please run it with the PyCharm. Note: you may first need to register an account on the offical website (https://www.eorc.jaxa.jp/) to obtain permission to download. Please ensure all purposes are for non-commercial.